[chores] Fixed selenium test imports
