+++ openwisp_controller/tests/utils.py
@@ -1,6 +1,5 @@
 from django.contrib.auth import get_user_model
 from django.urls import reverse
-from selenium.webdriver.common.by import By
 
 from openwisp_users.tests.utils import TestMultitenantAdminMixin
 
@@ -17,41 +16,3 @@ class TestAdminMixin(TestMultitenantAdminMixin):
 
     def _login(self, username='admin', password='tester'):
         self.client.force_login(user_model.objects.get(username=username))
-
-
-class SeleniumTestMixin:
-    """
-    A base test case for Selenium, providing helped methods for generating
-    clients and logging in profiles.
-    """
-
-    def open(self, url, driver=None):
-        """
-        Opens a URL
-        Argument:
-            url: URL to open
-            driver: selenium driver (default: cls.base_driver)
-        """
-        if not driver:
-            driver = self.web_driver
-        driver.get(f'{self.live_server_url}{url}')
-
-    def login(self, username=None, password=None, driver=None):
-        """
-        Log in to the admin dashboard
-        Argument:
-            driver: selenium driver (default: cls.web_driver)
-            username: username to be used for login (default: cls.admin.username)
-            password: password to be used for login (default: cls.admin.password)
-        """
-        if not driver:
-            driver = self.web_driver
-        if not username:
-            username = self.admin_username
-        if not password:
-            password = self.admin_password
-        driver.get(f'{self.live_server_url}/admin/login/')
-        if 'admin/login' in driver.current_url:
-            driver.find_element(by=By.NAME, value='username').send_keys(username)
-            driver.find_element(by=By.NAME, value='password').send_keys(password)
-            driver.find_element(by=By.XPATH, value='//input[@type="submit"]').click()
